+++ src/routes/+page.svelte
@@ -1,9 +1,15 @@
 <script lang="ts">
-  import * as m from '$lib/paraglide/messages.js'
+  import * as m from '$lib/paraglide/messages.js';
+
+
 </script>
 
-<h1>{m.title()}</h1>
-<!-- <p>Visit <a href="https://svelte.dev/docs/kit">svelte.dev/docs/kit</a> to read the documentation</p> -->
 
 
+<div class="flex w-full m-4">
 
+  <h1>{m.title()}</h1>
+  <p>Visit <a href="https://svelte.dev/docs/kit">svelte.dev/docs/kit</a> to read the documentation</p>
+
+
+</div>